+++ packages/slingshot-adapter-aws/src/index.ts
@@ -1,0 +1,3 @@
+export { s3 } from './s3.ts'
+
+export type { S3AdapterArgs } from './types.ts'