+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,30 @@
 # nextjs
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-js/slingshot@0.1.1
+  - @saas-js/slingshot-react@0.0.5
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-js/slingshot-react@0.0.4
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [d6da17e]
+- Updated dependencies [bd954ba]
+- Updated dependencies [bd954ba]
+  - @saas-js/slingshot@0.1.0
+  - @saas-js/slingshot-react@0.0.3
+
 ## 0.1.1
 
 ### Patch Changes