+++ website/CHANGELOG.md
@@ -1,5 +1,24 @@
 # website
 
+## 0.0.4
+
+### Patch Changes
+
+- @saas-js/slingshot-react@0.0.5
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-js/slingshot-react@0.0.4
+
+## 0.0.2
+
+### Patch Changes
+
+- @saas-js/slingshot-react@0.0.3
+
 ## 0.0.1
 
 ### Patch Changes